+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/mpgd/MUVT/MUVTStripFactory.java
@@ -1,14 +1,14 @@
-package org.jlab.detector.geant4.v2.URWT;
+package org.jlab.detector.geant4.v2.MPGD.MUVT;
 
 import org.jlab.detector.calib.utils.DatabaseConstantProvider;
-import org.jlab.detector.geant4.v2.mpgd.trapezoid.AbstractMPGDTrapezoidStripFactory;
+import org.jlab.detector.geant4.v2.MPGD.trapezoid.MPGDTrapezoidStripFactory;
 import org.jlab.detector.volume.Geant4Basic;
 
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * URWT strip factory.
+ * MUVT strip factory.
  *
  * It relies entirely on AbstractMPGDTrapezoidStripFactory for:
  *  - strip building (component IDs, endpoints)
@@ -18,37 +18,35 @@ import java.util.Map;
  * The ONLY detector-specific thing here is the mapping "volume name -> Geant4Basic",
  * used by the abstract class to find the sensitive volume (Sensitivity==1) and its transform.
  */
-public final class URWTStripFactory extends AbstractMPGDTrapezoidStripFactory {
+public final class MUVTStripFactory extends MPGDTrapezoidStripFactory {
 
     private final Map<String, Geant4Basic> volumeByName = new HashMap<>();
 
     /**
      * Build using an already-configured DatabaseConstantProvider.
-     * @param cp
-     * @param variation
      */
-    public URWTStripFactory(DatabaseConstantProvider cp, String variation) {
-        super(URWTConstants.getInstance(), new URWTGeant4Factory(cp, variation));
+public MUVTStripFactory(DatabaseConstantProvider cp, String variation) {
+    super(MUVTConstants.getInstance(), new MUVTGeant4Factory(cp, variation));
 
-        URWTConstants.connect(cp);
+    MUVTConstants.connect(cp);
 
-        for (Geant4Basic v : geo.getAllVolumes()) {
-            if (v.getName() != null) {
-                volumeByName.put(v.getName(), v);
-            }
+    for (Geant4Basic v : geo.getAllVolumes()) {
+        if (v.getName() != null) {
+            volumeByName.put(v.getName(), v);
         }
-
-        buildAll();
     }
+
+    buildAll();
+}
 
     /**
      * Convenience constructor: internally creates a DatabaseConstantProvider.
      */
-    public URWTStripFactory(String variation, int run) {
-        super(URWTConstants.getInstance(), new URWTGeant4Factory(variation, run));
+    public MUVTStripFactory(String variation, int run) {
+        super(MUVTConstants.getInstance(), new MUVTGeant4Factory(variation, run));
 
         DatabaseConstantProvider cp = new DatabaseConstantProvider(run, variation);
-        URWTConstants.connect(cp);
+        MUVTConstants.connect(cp);
         cp.disconnect();
 
         for (Geant4Basic v : geo.getAllVolumes()) {
@@ -65,14 +63,11 @@ public final class URWTStripFactory extends AbstractMPGDTrapezoidStripFactory {
         return volumeByName.get(name);
     }
 
-    /**
-     * If URWT uses 2D readout (same readout Z for both layers), enable this.
-     * Otherwise leave it false (default 1D: z = +/- zHalf).
-     */
+
     @Override
     protected boolean is2DReadout() {
-        // change to true if your URWT readout is truly 2D
-        return false;
+        // change to true if your MUVT readout is truly 2D
+        return true;
     }
 
     /**
@@ -81,19 +76,19 @@ public final class URWTStripFactory extends AbstractMPGDTrapezoidStripFactory {
     public static void main(String[] args) {
 
         int run = 11;
-        String variation = "urwt1";
+        String variation = "default";
 
         if (args.length > 0) {
             try { run = Integer.parseInt(args[0]); } catch (Exception ignored) {}
         }
         if (args.length > 1) variation = args[1];
 
-        URWTStripFactory sf = new URWTStripFactory(variation, run);
+        MUVTStripFactory sf = new MUVTStripFactory(variation, run);
 
-        int sector = 1;
-        int layer  = 4;
+        int sector = 2;
+        int layer  = 12;
 
-        System.out.println("URWT strips: sector=" + sector + " layer=" + layer
+        System.out.println("MUVT strips: sector=" + sector + " layer=" + layer
                 + " nComponents=" + sf.getNComponents(sector, layer));
 
         // print first strip global/local/tilted
