+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/mpgd/URWT/URWTConstants.java
@@ -1,25 +1,25 @@
-package org.jlab.detector.geant4.v2.MUVT;
+package org.jlab.detector.geant4.v2.MPGD.URWT;
 
 import org.jlab.detector.calib.utils.DatabaseConstantProvider;
-import org.jlab.detector.geant4.v2.mpgd.trapezoid.AbstractMPGDTrapezoidConstants;
+import org.jlab.detector.geant4.v2.MPGD.trapezoid.MPGDTrapezoidConstants;
 
 /**
  * URWT-specific constants.
  */
-public final class MUVTConstants extends AbstractMPGDTrapezoidConstants {
+public final class URWTConstants extends MPGDTrapezoidConstants {
 
 
-    private static final MUVTConstants INSTANCE =
-            new MUVTConstants(
-                    "/test/muvt/",       // CCDB base path
-                    "muvt_global",       // global table name
-                    "muvt_material"     // material table name
+    private static final URWTConstants INSTANCE =
+            new URWTConstants(
+                    "/test/urwt/",       // CCDB base path
+                    "urwt_global",       // global table name
+                    "urwt_material"  // material table name
             );
 
     /**
      * Private constructor: only the singleton instance is used.
      */
-    private MUVTConstants(String ccdbPath,
+    private URWTConstants(String ccdbPath,
                           String globalTable,
                           String materialTable) {
         super(ccdbPath, globalTable, materialTable);
@@ -29,7 +29,7 @@ public final class MUVTConstants extends AbstractMPGDTrapezoidConstants {
      * Returns the singleton instance for URWT constants.
      * @return 
      */
-    public static MUVTConstants getInstance() {
+    public static URWTConstants getInstance() {
         return INSTANCE;
     }
 
@@ -38,7 +38,7 @@ public final class MUVTConstants extends AbstractMPGDTrapezoidConstants {
      * {@link DatabaseConstantProvider}.Usage:
    DatabaseConstantProvider cp = new DatabaseConstantProvider(run, variation);
    URWTConstants.connect(cp);
-   // now MUVTConstants.getInstance() is fully initialized
+   // now URWTConstants.getInstance() is fully initialized
      *
      * @param cp
      * @return
