+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/mpgd/URWT/URWTGeant4Factory.java
@@ -1,19 +1,19 @@
-package org.jlab.detector.geant4.v2.URWT;
+package org.jlab.detector.geant4.v2.MPGD.URWT;
 
 import org.jlab.detector.calib.utils.DatabaseConstantProvider;
-import org.jlab.detector.geant4.v2.mpgd.trapezoid.AbstractMPGDTrapezoidGeant4Factory;
+import org.jlab.detector.geant4.v2.MPGD.trapezoid.MPGDTrapezoidGeant4Factory;
 
 /**
  * Geant4 factory for the uRWell Tracker (URWT).
  *
- * This class specializes the generic {@link AbstractMPGDTrapezoidGeant4Factory}
+ * This class specializes the generic {@link MPGDTrapezoidGeant4Factory}
  * by: - passing the URWT-specific constants - using "uRWT" as detector name in
  * volume names
  *
  * All the geometry construction (sectors, regions, material stack) is
  * implemented in the base class.
  */
-public final class URWTGeant4Factory extends AbstractMPGDTrapezoidGeant4Factory {
+public final class URWTGeant4Factory extends MPGDTrapezoidGeant4Factory {
 
     private final String variation;
