+++ common-tools/clas-reco/src/main/java/org/jlab/clas/reco/EngineProcessor.java
@@ -154,7 +154,7 @@ public class EngineProcessor {
             "org.jlab.rec.ft.hodo.FTHODOEngine",
             "org.jlab.rec.ft.trk.FTTRKEngine",
             "org.jlab.rec.ft.FTEBEngine",
-            "org.jlab.service.urwell.URWellEngine",
+            "org.jlab.service.urwt.URWTEngine",
             "org.jlab.service.dc.DCHBClustering",
             "org.jlab.service.dc.DCHBPostClusterConv",
             "org.jlab.service.ftof.FTOFHBEngine",
